Restructure + rewrite of a lot of stuff
Directly included mach_inject rather than referencing it with a
submodule. Rewrote parts of main.m to be a lot cleaner and more
understandable.

The injector is now a universal binary that can inject into any
process (it will silently relaunch itself under the hood in the
correct archiecture if needed). It also can pass arguments to
payloads.

Fixed #8.